Changing the the implementaiton of the extension point
