+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Base/Header/avatar/Avatar.jsx
@@ -11,7 +11,6 @@ import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import Configurations from 'Config';
-import AvatarMenu from './AvatarMenu';
 
 const styles = (theme) => ({
     profileMenu: {
@@ -125,18 +124,14 @@ class Avatar extends Component {
                     }}
                     className={classes.profileMenu}
                 >
-                    {AvatarMenu
-                        ? (<AvatarMenu doOIDCLogout={this.doOIDCLogout} />)
-                        : (
-                            <Link to={{ pathname: '/services/logout' }}>
-                                <MenuItem onClick={this.doOIDCLogout} id='logout'>
-                                    <FormattedMessage
-                                        id='Base.Header.avatar.Avatar.logout'
-                                        defaultMessage='Logout'
-                                    />
-                                </MenuItem>
-                            </Link>
-                        )}
+                    <Link to={{ pathname: '/services/logout' }}>
+                        <MenuItem onClick={this.doOIDCLogout} id='logout'>
+                            <FormattedMessage
+                                id='Base.Header.avatar.Avatar.logout'
+                                defaultMessage='Logout'
+                            />
+                        </MenuItem>
+                    </Link>
                 </Menu>
             </>
         );
