+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Base/Header/index.jsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,152 +16,27 @@
  * under the License.
  */
 import React from 'react';
-import { IconButton, Toolbar, AppBar } from '@material-ui/core';
-import { withStyles } from '@material-ui/core/styles';
-import MenuIcon from '@material-ui/icons/Menu';
-import SearchIcon from '@material-ui/icons/SearchOutlined';
-import Hidden from '@material-ui/core/Hidden';
 import PropTypes from 'prop-types';
-import { Link } from 'react-router-dom';
-import VerticalDivider from 'AppComponents/Shared/VerticalDivider';
-import RightTopMenu from 'AppComponents/Base/Header/rightTopMenu/RightTopMenu';
-import Configurations from 'Config';
-import Avatar from './avatar/Avatar';
-import HeaderSearch from './headersearch/HeaderSearch';
-import GlobalNavBar from './navbar/GlobalNavBar';
-
-const styles = (theme) => ({
-    appBar: {
-        position: 'relative',
-        background: theme.palette.background.appBar,
-    },
-    typoRoot: {
-        marginLeft: theme.spacing(3),
-        marginRight: theme.spacing(3),
-        textTransform: 'capitalize',
-    },
-    brandLink: {
-        color: theme.palette.primary.contrastText,
-    },
-    toolbar: {
-        minHeight: 56,
-        [`${theme.breakpoints.up('xs')} and (orientation: landscape)`]: {
-            minHeight: 48,
-        },
-        [theme.breakpoints.up('sm')]: {
-            minHeight: 64,
-        },
-    },
-    menuIcon: {
-        color: theme.palette.getContrastText(theme.palette.background.appBar),
-        fontSize: 35,
-    },
-});
+import HeaderOriginal from './HeaderOriginal';
 
 /**
- * Construct the Global AppBar header section
- * @class Header
- * @extends {React.Component}
+ * Acts as an extension point.
+ * You can return the Header original component with additional
+ * props to override the defualt look and feel.
+ * <HeaderOriginal avatar={<NewSettings />}
+        user={user}
+        settings={<NewSettings />}
+        menuItems={[<SettingsButton />, <SettingsButton />]}
+    />
+ * @param {object} props props passed down from the parent.
+ * @returns {element} Original header component.
  */
-class Header extends React.Component {
-    /**
-     *Creates an instance of Header.
-     * @param {Object} props @inheritdoc
-     * @memberof Header
-     */
-    constructor(props) {
-        super(props);
-        this.state = {
-            openNavBar: false,
-            smScreen: false,
-        };
-        this.toggleGlobalNavBar = this.toggleGlobalNavBar.bind(this);
-        this.toggleSmSearch = this.toggleSmSearch.bind(this);
-    }
-
-    /**
-     * Toggle the Global LHS Navbar visibility
-     *
-     * @memberof Header
-     */
-    toggleGlobalNavBar() {
-        const { openNavBar } = this.state;
-        this.setState({ openNavBar: !openNavBar });
-    }
-
-    /**
-     * Show search input in sm breakpoint or lower resolution
-     */
-    toggleSmSearch() {
-        const { smScreen } = this.state;
-        this.setState({ smScreen: !smScreen });
-    }
-
-    /**
-     *
-     * @inheritdoc
-     * @returns {React.ComponentClass} @inheritdoc
-     * @memberof Header
-     */
-    render() {
-        const { openNavBar, smScreen } = this.state;
-        const {
-            classes, avatar, settings, theme,
-        } = this.props;
-        return (
-            <>
-                <AppBar className={classes.appBar} position='fixed'>
-                    <Toolbar className={classes.toolbar}>
-                        <Hidden mdUp>
-                            <IconButton onClick={this.toggleGlobalNavBar}>
-                                <MenuIcon className={classes.menuIcon} />
-                            </IconButton>
-                        </Hidden>
-                        <Link to='/'>
-                            <img
-                                src={Configurations.app.context + theme.custom.logo}
-                                alt={theme.custom.title}
-                                style={{ height: theme.custom.logoHeight, width: theme.custom.logoWidth }}
-                            />
-                        </Link>
-                        <GlobalNavBar toggleGlobalNavBar={this.toggleGlobalNavBar} open={openNavBar} />
-                        <VerticalDivider height={32} />
-                        <Hidden smDown>
-                            <HeaderSearch />
-                        </Hidden>
-                        <Hidden mdUp>
-                            <IconButton onClick={this.toggleSmSearch} color='inherit'>
-                                <SearchIcon className={classes.menuIcon} />
-                            </IconButton>
-                            {smScreen && <HeaderSearch toggleSmSearch={this.toggleSmSearch} smSearch={smScreen} />}
-                        </Hidden>
-                        {settings}
-                        {avatar}
-                    </Toolbar>
-                </AppBar>
-            </>
-        );
-    }
+export default function Header(props) {
+    const { avatar, user } = props;
+    return <HeaderOriginal avatar={avatar} user={user} />;
 }
-Header.defaultProps = {
-    avatar: <Avatar />,
-    settings: <RightTopMenu />,
-};
 
 Header.propTypes = {
-    classes: PropTypes.shape({
-        appBar: PropTypes.string,
-        menuIcon: PropTypes.string,
-        toolbar: PropTypes.string,
-    }).isRequired,
-    avatar: PropTypes.element,
-    settings: PropTypes.element,
-    theme: PropTypes.shape({
-        custom: PropTypes.shape({
-            logo: PropTypes.string,
-            title: PropTypes.string,
-        }),
-    }).isRequired,
+    avatar: PropTypes.element.isRequired,
+    user: PropTypes.shape({ name: PropTypes.string.isRequired }).isRequired,
 };
-
-export default withStyles(styles, { withTheme: true })(Header);
