+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Base/Header/settings/SettingsButton.jsx
@@ -20,7 +20,6 @@ import { Icon, IconButton, withStyles } from '@material-ui/core';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import { FormattedMessage } from 'react-intl';
-import RightTopMenuLinks from './RightTopMenuLinks';
 
 const styles = (theme) => ({
     settingsIconbtn: {
@@ -36,11 +35,6 @@ const styles = (theme) => ({
 
 const SettingsButton = (props) => {
     const { classes } = props;
-    if (RightTopMenuLinks) {
-        return (
-            <RightTopMenuLinks />
-        );
-    }
     return (
         <>
             <Link to='/settings'>
